refactor: update ZapButton imports and restructure settings management across components
