+++ core/src/lib/settings.ts
