+++ core/src/zap-old/components/waitlist/waitlist-form.tsx
@@ -10,8 +10,8 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { ZapButton } from "@/components/zap-ui/button";
 import { cn } from "@/lib/utils";
+import { ZapButton } from "@/zap/components/button";
 import { useWaitlist } from "@/zap/hooks/waitlist/use-waitlist";
 
 export function WaitlistForm() {
