+++ core/src/zap/auth/components/forms/register-form.tsx
@@ -13,9 +13,9 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { ZapButton } from "@/components/zap-ui/button";
 import { useAuth } from "@/zap/auth/hooks";
 import { RegisterFormSchema } from "@/zap/auth/schemas";
+import { ZapButton } from "@/zap/components/button";
 
 type RegisterFormValues = z.infer<typeof RegisterFormSchema>;
