+++ core/src/zap/auth/components/index.ts
@@ -1,3 +1,3 @@
-export * from "@/zap/auth/components/auth-page";
+export * from "@/zap/auth/components/auth-icons";
 export * from "@/zap/auth/components/forms";
 export * from "@/zap/auth/components/social-provider-button";
