+++ core/src/zap/auth/pages/index.ts
@@ -1,0 +1,1 @@
+export * from "@/zap/auth/pages/auth.page";
