+++ core/src/zap/legal/pages/index.ts
@@ -1,0 +1,1 @@
+export * from "@/zap/legal/pages/legal.page";
